Provide a fixtureFile test helper
* The mocked "wicked show-xml" and "wicked show-config" are now outside
  of __mocks__.
